Extracted some code from act_comm.cpp
