+++ src/act_comm.cpp
@@ -24,11 +24,6 @@
 #include "gameplay/communication/ignores.h"
 #include "engine/ui/modify.h"
 
-// local functions
-void perform_tell(CharData *ch, CharData *vict, char *arg);
-int is_tell_ok(CharData *ch, CharData *vict);
-bool tell_can_see(CharData *ch, CharData *vict);
-
 // external functions
 extern char *diag_timer_to_char(const ObjData *obj);
 extern void SetWait(CharData *ch, int waittime, int victim_in_room);
@@ -44,179 +39,6 @@ void do_gen_comm(CharData *ch, char *argument, int cmd, int subcmd);
 void do_pray_gods(CharData *ch, char *argument, int cmd, int subcmd);
 void do_ignore(CharData *ch, char *argument, int cmd, int subcmd);
 
-void perform_tell(CharData *ch, CharData *vict, char *arg) {
-	if (vict->IsFlagged(EPrf::kNoInvistell)
-		&& !CAN_SEE(vict, ch)
-		&& GetRealLevel(ch) < kLvlImmortal
-		&& !ch->IsFlagged(EPrf::kCoderinfo)) {
-		act("$N не любит разговаривать с теми, кого не видит.", false, ch, nullptr, vict, kToChar | kToSleep);
-		return;
-	}
-
-	// TODO: если в act() останется показ иммов, то это и эхо ниже переделать на act()
-	if (tell_can_see(ch, vict)) {
-		snprintf(buf, kMaxStringLength, "%s сказал%s вам : '%s'", GET_NAME(ch), GET_CH_SUF_1(ch), arg);
-	} else {
-		snprintf(buf, kMaxStringLength, "Кто-то сказал вам : '%s'", arg);
-	}
-	snprintf(buf1, kMaxStringLength, "%s%s%s\r\n", kColorBoldCyn, CAP(buf), kColorNrm);
-	SendMsgToChar(buf1, vict);
-	if (!vict->IsNpc()) {
-		vict->remember_add(buf1, Remember::ALL);
-	}
-
-	if (!vict->IsNpc() && !ch->IsNpc()) {
-		snprintf(buf, kMaxStringLength, "%s%s : '%s'%s\r\n", kColorBoldCyn,
-				 tell_can_see(ch, vict) ? GET_NAME(ch) : "Кто-то", arg, kColorNrm);
-		vict->remember_add(buf, Remember::PERSONAL);
-	}
-
-	if (ch->IsFlagged(EPrf::kNoRepeat)) {
-		SendMsgToChar(OK, ch);
-	} else {
-		snprintf(buf, kMaxStringLength, "%sВы сказали %s : '%s'%s\r\n", kColorBoldCyn,
-				 tell_can_see(vict, ch) ? vict->player_data.PNames[2].c_str() : "кому-то", arg, kColorNrm);
-		SendMsgToChar(buf, ch);
-		if (!ch->IsNpc()) {
-			ch->remember_add(buf, Remember::ALL);
-		}
-	}
-
-	if (!vict->IsNpc() && !ch->IsNpc()) {
-		vict->set_answer_id(GET_UID(ch));
-	}
-}
-
-int is_tell_ok(CharData *ch, CharData *vict) {
-	if (ch == vict) {
-		SendMsgToChar("Вы начали потихоньку разговаривать с самим собой.\r\n", ch);
-		return (false);
-	} else if (ch->IsFlagged(EPlrFlag::kDumbed)) {
-		SendMsgToChar("Вам запрещено обращаться к другим игрокам.\r\n", ch);
-		return (false);
-	} else if (!vict->IsNpc() && !vict->desc)    // linkless
-	{
-		act("$N потерял$G связь в этот момент.", false, ch, nullptr, vict, kToChar | kToSleep);
-		return (false);
-	} else if (vict->IsFlagged(EPlrFlag::kWriting)) {
-		act("$N пишет сообщение - повторите попозже.", false, ch, nullptr, vict, kToChar | kToSleep);
-		return (false);
-	}
-
-	if (IS_GOD(ch) || ch->IsFlagged(EPrf::kCoderinfo))
-		return (true);
-
-	if (ROOM_FLAGGED(ch->in_room, ERoomFlag::kSoundproof))
-		SendMsgToChar(SOUNDPROOF, ch);
-	else if ((!vict->IsNpc() &&
-		(vict->IsFlagged(EPrf::kNoTell) || ignores(vict, ch, EIgnore::kTell))) ||
-		ROOM_FLAGGED(vict->in_room, ERoomFlag::kSoundproof)) {
-		if (ch->IsNpc()) {
-			if (ROOM_FLAGGED(vict->in_room, ERoomFlag::kSoundproof)) {
-				log("NOTELL: name %s room %d soundproof", GET_NAME(vict), GET_ROOM_VNUM(vict->in_room));
-			}
-			if (vict->IsFlagged(EPrf::kNoTell))
-				log("NOTELL: name %s глух как пробка", GET_NAME(vict));
-		}
-		act("$N не сможет вас услышать.", false, ch, nullptr, vict, kToChar | kToSleep);
-	} else if (vict->GetPosition() < EPosition::kRest || AFF_FLAGGED(vict, EAffect::kDeafness))
-		act("$N вас не услышит.", false, ch, nullptr, vict, kToChar | kToSleep);
-	else
-		return (true);
-
-	return (false);
-}
-
-/*
- * Yes, do_tell probably could be combined with whisper and ask, but
- * called frequently, and should IMHO be kept as tight as possible.
- */
-void do_tell(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-	CharData *vict = nullptr;
-
-	if (AFF_FLAGGED(ch, EAffect::kCharmed))
-		return;
-
-	if (AFF_FLAGGED(ch, EAffect::kSilence)) {
-		SendMsgToChar(SIELENCE, ch);
-		return;
-	}
-
-	/* Непонятно нафига нужно
-	if (ROOM_FLAGGED(ch->in_room, ERoomFlag::kTribune))
-	{
-		SendMsgToChar(SOUNDPROOF, ch);
-		return;
-	}
-	*/
-
-	half_chop(argument, buf, buf2);
-
-	if (!*buf || !*buf2) {
-		SendMsgToChar("Что и кому вы хотите сказать?\r\n", ch);
-	} else if (!(vict = get_player_vis(ch, buf, EFind::kCharInWorld))) {
-		SendMsgToChar(NOPERSON, ch);
-	} else if (vict->IsNpc())
-		SendMsgToChar(NOPERSON, ch);
-	else if (is_tell_ok(ch, vict)) {
-		if (ch->IsFlagged(EPrf::kNoTell))
-			SendMsgToChar("Ответить вам не смогут!\r\n", ch);
-		perform_tell(ch, vict, buf2);
-	}
-}
-
-void do_reply(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-	if (ch->IsNpc())
-		return;
-
-	if (AFF_FLAGGED(ch, EAffect::kSilence)) {
-		SendMsgToChar(SIELENCE, ch);
-		return;
-	}
-
-	if (ch->IsFlagged(EPlrFlag::kDumbed)) {
-		SendMsgToChar("Вам запрещено обращаться к другим игрокам!\r\n", ch);
-		return;
-	}
-
-	skip_spaces(&argument);
-
-	if (ch->get_answer_id() == kNobody)
-		SendMsgToChar("Вам некому ответить!\r\n", ch);
-	else if (!*argument)
-		SendMsgToChar("Что вы собираетесь ответить?\r\n", ch);
-	else {
-		/*
-		 * Make sure the person you're replying to is still playing by searching
-		 * for them.  Note, now last tell is stored as player IDnum instead of
-		 * a pointer, which is much better because it's safer, plus will still
-		 * work if someone logs out and back in again.
-		 */
-
-		/*
-		 * XXX: A descriptor list based search would be faster although
-		 *      we could not find link dead people.  Not that they can
-		 *      hear tells anyway. :) -gg 2/24/98
-		 */
-		bool found = false;
-		for (const auto &i : character_list) {
-			if (!i->IsNpc()
-				&& GET_UID(i) == ch->get_answer_id()) {
-				if (is_tell_ok(ch, i.get())) {
-					perform_tell(ch, i.get(), argument);
-				}
-
-				found = true;
-
-				break;
-			}
-		}
-
-		if (!found) {
-			SendMsgToChar("Этого игрока уже нет в игре.\r\n", ch);
-		}
-	}
-}
 
 void do_spec_comm(CharData *ch, char *argument, int/* cmd*/, int subcmd) {
 	CharData *vict;
@@ -674,65 +496,4 @@ void do_pray_gods(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
 	}
 }
 
-/**
-* Канал оффтоп. Не виден иммам, всегда видно кто говорит, вкл/выкл режим оффтоп.
-*/
-void do_offtop(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-	if (ch->IsNpc() || GetRealLevel(ch) >= kLvlImmortal || ch->IsFlagged(EPrf::kStopOfftop)) {
-		SendMsgToChar("Чаво?\r\n", ch);
-		return;
-	}
 
-	if (ch->IsFlagged(EPlrFlag::kDumbed) || ch->IsFlagged(EPlrFlag::kMuted)) {
-		SendMsgToChar("Вам запрещено обращаться к другим игрокам!\r\n", ch);
-		return;
-	}
-	//if (ROOM_FLAGGED(ch->in_room, ERoomFlag::kSoundproof) || ROOM_FLAGGED(ch->in_room, ERoomFlag::kTribune))
-	if (ROOM_FLAGGED(ch->in_room, ERoomFlag::kSoundproof)) {
-		SendMsgToChar(SOUNDPROOF, ch);
-		return;
-	}
-	if (GetRealLevel(ch) < offtop_system::kMinOfftopLvl && !GetRealRemort(ch)) {
-		SendMsgToChar(ch, "Вам стоит достичь хотя бы %d уровня, чтобы вы могли оффтопить.\r\n",
-					  offtop_system::kMinOfftopLvl);
-		return;
-	}
-	if (!ch->IsFlagged(EPrf::kOfftopMode)) {
-		SendMsgToChar("Вы вне видимости канала.\r\n", ch);
-		return;
-	}
-	skip_spaces(&argument);
-	if (!*argument) {
-		SendMsgToChar(ch, "Раз нечего сказать, то лучше уж помолчать.");
-		return;
-	}
-	utils::ConvertToLow(argument);
-	if (!strcmp(ch->get_last_tell().c_str(), argument)) {
-		SendMsgToChar("Но вы же недавно говорили тоже самое!?!\r\n", ch);
-		return;
-	}
-	// эта проверка должна идти последней и послее нее мессага полюбому идет в эфир
-	if (!antispam::check(ch, antispam::kOfftopMode)) {
-		return;
-	}
-	ch->set_last_tell(argument);
-	if (ch->IsFlagged(EPlrFlag::kSpamer)) // а вот фиг, еще проверка :)
-		return;
-	snprintf(buf, kMaxStringLength, "[оффтоп] %s : '%s'\r\n", GET_NAME(ch), argument);
-	snprintf(buf1, kMaxStringLength, "&c%s&n", buf);
-	for (DescriptorData *i = descriptor_list; i; i = i->next) {
-		// переплут как любитель почитывать логи за ночь очень хотел этот канал...
-		// а мы шо, не люди? даешь оффтоп 34-ым! кому не нравится - реж оффтоп...
-		if (STATE(i) == CON_PLAYING
-			&& i->character
-// оффтоп видят все, временно
-//			&& (GetRealLevel(i->character) < kLvlImmortal || IS_IMPL(i->character))
-			&& i->character->IsFlagged(EPrf::kOfftopMode)
-			&& !i->character->IsFlagged(EPrf::kStopOfftop)
-			&& !ignores(i->character.get(), ch, EIgnore::kOfftop)) {
-			SendMsgToChar(i->character.get(), "%s%s%s", kColorCyn, buf, kColorNrm);
-			i->character->remember_add(buf1, Remember::ALL);
-		}
-	}
-	ch->remember_add(buf1, Remember::OFFTOP);
-}
