+++ src/engine/ui/cmd/do_tell.h
@@ -1,0 +1,17 @@
+/**
+\file do_tell.h - a part of the Bylins engine.
+\authors Created by Sventovit.
+\date 23.09.2024.
+\brief Brief description.
+\detail Detail description.
+*/
+
+#ifndef BYLINS_SRC_ENGINE_UI_CMD_DO_TELL_H_
+#define BYLINS_SRC_ENGINE_UI_CMD_DO_TELL_H_
+
+class CharData;
+void do_tell(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_ENGINE_UI_CMD_DO_TELL_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
